Create solve.rs
